Removed binary files, and expanded the .gitignore file.
This commit removes excess binary files, which Git is not designed for,
and additionally extends the `.gitignore` file to prevent such files
from being added. Such files tend to take up immense space within the
repository over time, cannot be diffed properly, and are prone to being
tampered with by threat actors.

This commit is a security improvement, and will prevent the Git
repository from bloating in the future